+++ client/src/components/FormSignIn/FormSignIn.module.css
@@ -1,0 +1,5 @@
+.signinForm__input {
+  color: red;
+  font-size: 40px;
+  padding: 20px;
+}
