--- client/src/pages/homePage/HomePage.module.css
@@ -1,3 +1,0 @@
-div {
-  color: red;
-}
